+++ backend/AAiT-backend-group-5/Config/config.go
@@ -1,0 +1,1 @@
+package config
