+++ backend/AAiT-backend-group-5/Domain/Interfaces/blog_repository_interface.go
@@ -1,0 +1,1 @@
+package interfaces
