+++ backend/AAiT-backend-group-5/Domain/Interfaces/blog_usecase_interface.go
@@ -1,0 +1,1 @@
+package interfaces
