+++ backend/AAiT-backend-group-5/Domain/Models/session_model.go
@@ -1,0 +1,10 @@
+package models
+
+import "go.mongodb.org/mongo-driver/bson/primitive"
+
+type Session struct {
+	ID           primitive.ObjectID `bson:"_id,omitempty" json:"id"`
+	UserID       primitive.ObjectID `bson:"user_id" json:"user_id"`
+	AccessToken  string             `bson:"access_token" json:"access_token"`
+	RefreshToken string             `bson:"refresh_token" json:"refresh_token"`
+}
