+++ backend/AAiT-backend-group-5/UseCases/blog_usecases.go
@@ -1,0 +1,1 @@
+package usecases
