+++ backend/AAiT-backend-group-5/UseCases/user_usecases.go
@@ -1,0 +1,1 @@
+package usecases
