+++ backend/AAiT-backend-group-5/cmd/main.go
@@ -1,0 +1,1 @@
+package main
